# Improve AVX2 hash speed.