--- a/avo/avx.go
+++ b/avo/avx.go
@@ -7,9 +7,11 @@ import (
 )
 
 func AVX() {
-	// Lay out the prime constant in memory
+	// Lay out the prime constant in memory, copy it so no unpack is needed.
 	primeData := GLOBL("prime_avx", RODATA|NOPTR)
-	DATA(0, U32(2654435761))
+	for i := 0; i < 32; i += 8 {
+		DATA(i, U32(2654435761))
+	}
 
 	{
 		TEXT("accumAVX2", NOSPLIT, "func(acc *[8]uint64, data, key *byte, len uint64)")
@@ -32,7 +34,7 @@ func AVX() {
 		{
 			VMOVDQU(acc.Offset(0x00), a[0])
 			VMOVDQU(acc.Offset(0x20), a[1])
-			VPBROADCASTQ(primeData, prime)
+			VMOVDQU(primeData, prime)
 		}
 
 		Label("accum_large")
@@ -41,7 +43,7 @@ func AVX() {
 			JLE(LabelRef("accum"))
 
 			for i := 0; i < 16; i++ {
-				avx2accum(data, a, 64*i, 8*i, key)
+				avx2accum(data, a, 64*i, 8*i, key, true)
 			}
 			advance(16)
 			avx2scramble(prime, key, a, 8*16)
@@ -54,7 +56,7 @@ func AVX() {
 			CMPQ(plen, Imm(64))
 			JLE(LabelRef("finalize"))
 
-			avx2accum(data, a, 0, 0, skey)
+			avx2accum(data, a, 0, 0, skey, false)
 			advance(1)
 			ADDQ(U32(8), skey.Base)
 
@@ -69,7 +71,7 @@ func AVX() {
 			SUBQ(Imm(64), data.Base)
 			ADDQ(plen, data.Base)
 
-			avx2accum(data, a, 0, 121, key)
+			avx2accum(data, a, 0, 121, key, false)
 		}
 
 		Label("return")
@@ -94,13 +96,13 @@ func AVX() {
 		{
 			VMOVDQU(acc.Offset(0x00), a[0])
 			VMOVDQU(acc.Offset(0x20), a[1])
-			VPBROADCASTQ(primeData, prime)
+			VMOVDQU(primeData, prime)
 		}
 
 		Label("accum_block")
 		{
 			for i := 0; i < 16; i++ {
-				avx2accum(data, a, 64*i, 8*i, key)
+				avx2accum(data, a, 64*i, 8*i, key, false)
 			}
 			avx2scramble(prime, key, a, 8*16)
 		}
@@ -133,17 +135,24 @@ func avx2scramble(prime VecVirtual, key Mem, a [2]VecVirtual, koff int) {
 	}
 }
 
-func avx2accum(data Mem, a [2]VecVirtual, doff, koff int, key Mem) {
-	for n, offset := range []int{0x00, 0x20} {
-		y0, y1, y2 := YMM(), YMM(), YMM()
-
-		VMOVDQU(data.Offset(doff+offset), y0)
-		VMOVDQU(key.Offset(koff+offset), y1)
-		VPXOR(y0, y1, y1)
-		VPSHUFD(Imm(49), y1, y2)
-		VPMULUDQ(y1, y2, y1)
-		VPSHUFD(Imm(78), y0, y0)
-		VPADDQ(a[n], y0, a[n])
-		VPADDQ(a[n], y1, a[n])
+func avx2accum(data Mem, a [2]VecVirtual, doff, koff int, key Mem, prefetch bool) {
+	y0, y1, y2, y10, y11, y12 := YMM(), YMM(), YMM(), YMM(), YMM(), YMM()
+	VMOVDQU(data.Offset(doff+0x00), y0)
+	VMOVDQU(data.Offset(doff+0x20), y10)
+	if prefetch {
+		// Prefetch once per cacheline (64 bytes), 8 iterations ahead.
+		PREFETCHT0(data.Offset(doff + 512))
 	}
+	VPXOR(key.Offset(koff+0x00), y0, y1)
+	VPXOR(key.Offset(koff+0x20), y10, y11)
+	VPSHUFD(Imm(49), y1, y2)
+	VPSHUFD(Imm(49), y11, y12)
+	VPMULUDQ(y1, y2, y1)
+	VPMULUDQ(y11, y12, y11)
+	VPSHUFD(Imm(78), y0, y0)
+	VPSHUFD(Imm(78), y10, y10)
+	VPADDQ(a[0], y0, a[0])
+	VPADDQ(a[0], y1, a[0])
+	VPADDQ(a[1], y10, a[1])
+	VPADDQ(a[1], y11, a[1])
 }