… accum_vector_avx_amd64.s
… accum_vector_sse_amd64.s
… hash64_test.go