Toyed a bit around and it seems like the biggest change came from loading the key value directly into the XOR.

Maybe an AMD specific optimization.

```
λ go test -bench=/AVX2 -test.count=10

λ benchstat before.txt after.txt
name                     old time/op    new time/op    delta
Fixed128/241-AVX2-32       23.7ns ± 1%    23.7ns ± 1%    ~     (p=0.108 n=10+10)
Fixed128/512-AVX2-32       27.6ns ± 1%    27.5ns ± 1%  -0.44%  (p=0.050 n=10+10)
Fixed128/1024-AVX2-32      37.4ns ± 1%    35.7ns ± 1%  -4.63%  (p=0.000 n=10+10)
Fixed128/8192-AVX2-32       154ns ± 0%     144ns ± 0%  -6.34%  (p=0.000 n=10+10)
Fixed128/102400-AVX2-32    1.77µs ± 1%    1.64µs ± 1%  -7.48%  (p=0.000 n=10+10)
Fixed/241-AVX2-32          20.3ns ± 1%    20.1ns ± 0%  -1.11%  (p=0.000 n=10+10)
Fixed/512-AVX2-32          23.0ns ± 1%    23.4ns ± 0%  +1.41%  (p=0.000 n=10+10)
Fixed/1024-AVX2-32         33.2ns ± 1%    30.5ns ± 0%  -7.92%   (p=0.000 n=10+9)
Fixed/8192-AVX2-32          149ns ± 0%     138ns ± 1%  -7.44%   (p=0.000 n=9+10)
Fixed/102400-AVX2-32       1.76µs ± 1%    1.63µs ± 0%  -7.73%   (p=0.000 n=10+8)

name                     old speed      new speed      delta
Fixed128/241-AVX2-32     10.2GB/s ± 1%  10.2GB/s ± 1%    ~     (p=0.105 n=10+10)
Fixed128/512-AVX2-32     18.5GB/s ± 1%  18.6GB/s ± 1%    ~     (p=0.052 n=10+10)
Fixed128/1024-AVX2-32    27.4GB/s ± 1%  28.7GB/s ± 1%  +4.85%  (p=0.000 n=10+10)
Fixed128/8192-AVX2-32    53.3GB/s ± 0%  57.0GB/s ± 0%  +6.78%  (p=0.000 n=10+10)
Fixed128/102400-AVX2-32  57.9GB/s ± 1%  62.5GB/s ± 0%  +8.08%  (p=0.000 n=10+10)
Fixed/241-AVX2-32        11.9GB/s ± 1%  12.0GB/s ± 0%  +1.11%  (p=0.000 n=10+10)
Fixed/512-AVX2-32        22.2GB/s ± 1%  21.9GB/s ± 0%  -1.39%  (p=0.000 n=10+10)
Fixed/1024-AVX2-32       30.9GB/s ± 1%  33.5GB/s ± 0%  +8.55%  (p=0.000 n=10+10)
Fixed/8192-AVX2-32       54.8GB/s ± 0%  59.2GB/s ± 1%  +8.05%   (p=0.000 n=9+10)
Fixed/102400-AVX2-32     58.0GB/s ± 1%  62.9GB/s ± 0%  +8.38%   (p=0.000 n=10+8)
```

cpu: AMD Ryzen 9 3950X 16-Core Processor

Moved the generation into the avo folder, since I couldn't make the other work.