@klauspost 2021-08-30
> apparently, throughput is better when it's lower 

Yeah, that's how I read it as well. 

> I don't know why the loop version runs faster when the unrolled version gets slower, though. I'll play around with some different orderings.

This is the first that moves beyond L1 size, so maybe some prefetching isn't that optimal on the Intel.

I had suspected that for any superscalar CPU the rolled/unrolled would be the same, but was surprised when it wasn't.