@klauspost 2021-08-30
@zeebo Tweaked some more stuff after rebasing. Experimented a bit. Found that for once prefetching was faster.

`VZEROUPPER` was missing. Huge improvement on small payloads.

Top benchmark updated.