@zeebo 2021-08-30
I saw your comment in the gophers slack to remember to add `VZEROUPPER`, so I went and did that without first seeing this PR. :smile: 

So I rebased and fixed the conflict. This is great! I also saw improvements on the order of ~12% for the 100K payloads. Thanks!