@klauspost 2021-08-30
@zeebo I saw you <3 there, so I assumed you had seen the PR already :D

Great it improves Intel as well 👍🏼